function_files/requirements.txt
google-cloud-bigquery==2.27.1
google-cloud-logging==3.0.0
google-cloud-storage==1.42.2
pandas==1.3.3
requests==2.26.0
pyarrow==5.0.0